backend/requirements.txt
beautifulsoup4
requests
python-dotenv
fastapi[standard]
uvicorn[standard]